+++ src/main/java/net/sirplop/aetherworks/api/damage/EffectDamagePotion.java
@@ -10,10 +10,8 @@ import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.Attributes;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.phys.Vec3;
-import net.sirplop.aetherworks.Aetherworks;
 
 import javax.annotation.Nullable;
 import java.util.List;