+++ src/main/java/net/sirplop/aetherworks/effect/PullDownEffect.java
@@ -1,0 +1,41 @@
+package net.sirplop.aetherworks.effect;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectCategory;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+
+import java.awt.*;
+
+public class PullDownEffect extends MobEffect {
+
+    public static final Color COLOR = new Color(157, 98, 51);
+
+    public PullDownEffect(MobEffectCategory pCategory, int pColor) {
+        super(pCategory, pColor);
+    }
+
+    @Override
+    public void applyEffectTick(LivingEntity pLivingEntity, int pAmplifier) {
+        if (!pLivingEntity.level().isClientSide() || isClientPlayer(pLivingEntity)) {
+            //pLivingEntity.push(0, -0.25 * (1 + pAmplifier), 0);
+            pLivingEntity.push(0, -0.1f * (1f + pAmplifier), 0);
+        }
+
+        super.applyEffectTick(pLivingEntity, pAmplifier);
+    }
+
+    @OnlyIn(Dist.CLIENT)
+    private boolean isClientPlayer(LivingEntity livingEntity) {
+        return livingEntity instanceof Player player && Minecraft.getInstance().player == player;
+    }
+
+    @Override
+    public boolean isDurationEffectTick(int pDuration, int pAmplifier) {
+        return true;
+    }
+}